Adjust annotation handling in swagger view generation
The flask_swagger_generator has been updated to handle tuple types in schema creation from class annotation. If the provided annotation is a tuple, only the first element of the tuple will be considered. This ensures that no undesired characteristics are added to the schema, thus improving the accuracy and reliability of the generated Swagger view.